ProdStories/hw5/requirements.txt
torch
ray[rllib]~=1.5.2
ray[default]
gym~=0.21.0
numpy~=1.21.3
wandb~=0.12.5
PyYAML~=6.0
Pillow~=8.4.0